Refactor, move code in ShellCommand
