+++ client/src/main/java/io/cloudsoft/winrm4j/client/CliClient.java
@@ -6,12 +6,6 @@ import java.io.PrintWriter;
 public class CliClient {
 
     public static void main(String[] args) throws Exception{
-//        System.setProperty("com.sun.xml.ws.transport.http.client.HttpTransportPipe.dump", "true");
-//        System.setProperty("com.sun.xml.internal.ws.transport.http.client.HttpTransportPipe.dump", "true");
-//        System.setProperty("com.sun.xml.ws.transport.http.HttpAdapter.dump", "true");
-//        System.setProperty("com.sun.xml.internal.ws.transport.http.HttpAdapter.dump", "true");
-//        System.setProperty("com.sun.xml.internal.ws.transport.http.HttpAdapter.dumpTreshold", Integer.toString(Integer.MAX_VALUE));
-
         if (args.length != 4) {
             System.out.println("Usage: CliClient <endpoint> <username> <password> <command>");
         }
@@ -29,8 +23,8 @@ public class CliClient {
 //                .environment(env)
                 .build();
         int exitCode = 999;
-        try {
-            exitCode = client.command(cmd, new PrintWriter(new OutputStreamWriter(System.out)), new PrintWriter(new OutputStreamWriter(System.err)));
+        try (ShellCommand shell = client.createShell()){
+            exitCode = shell.execute(cmd, new PrintWriter(new OutputStreamWriter(System.out)), new PrintWriter(new OutputStreamWriter(System.err)));
         } finally {
             client.disconnect();
         }
